requirements.txt
streamlit>=1.30.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
google-generativeai>=0.4.0
pandas>=2.0.0
plotly>=5.18.0
supabase>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
seaborn>=0.13.0
numpy>=1.24.0
scipy>=1.11.0
streamlit-chat>=0.1.1